+++ src/commands/services.h
@@ -1,0 +1,7 @@
+#ifndef CMD_SERVICES_H
+#define CMD_SERVICES_H
+
+void cmd_chanserv(const char *);
+void cmd_nickserv(const char *);
+
+#endif